rename chocolates and users factory
